+++ src/modules/users/types/users.interface.ts
@@ -1,0 +1,7 @@
+export interface UserResponse {
+  id: string;
+  login: string;
+  version: number;
+  createdAt: Date;
+  updatedAt: Date;
+}
