--- src/types/artists.interface.ts
@@ -1,5 +1,0 @@
-export interface Artist {
-  id: string;
-  name: string;
-  grammy: boolean;
-}
