--- src/types/favorites.interface.ts
@@ -1,15 +1,0 @@
-import { Artist } from '@prisma/client';
-import { Album } from '@prisma/client';
-import { Track } from '@prisma/client';
-
-export interface Favorites {
-  artists: string[];
-  albums: string[];
-  tracks: string[];
-}
-
-export interface FavoritesResponse {
-  artists: Artist[];
-  albums: Album[];
-  tracks: Track[];
-}
