--- src/types/tracks.interface.ts
@@ -1,7 +1,0 @@
-export interface ITrack {
-  id: string;
-  name: string;
-  artistId: string | null;
-  albumId: string | null;
-  duration: number;
-}
